Subsets the mapped KO table based on pathways.

reads the output of the mapping_ko function to filter out KOs based on pathways.

get_subset_pathway(mapped_ko_table, type_of_interest_feature,
interest_feature)

Arguments

mapped_ko_table
a data frame object. The output table from the mapping_ko function.

type_of_interest_feature
is a mapped_ko column name. This feature is going to be used for subsetting.

interest_feature
a character vector of the mapped_ko feature of interest. It is found under the type_of_interest_feature value column.

Details

This function is part of a package used for the analysis of bins metabolism.

Examples

get_subset_pathway(mapped_ko_table=ko_bin_mapp, 
type_of_interest_feature=rbims_pathway, interest_feature="Hexadecane")

#> # A tibble: 0 x 19
#> # … with 19 variables: Module <chr>, Module_description <chr>, Pathway <chr>,
#> #   Pathway_description <chr>, Cycle <chr>, Pathway_cycle <chr>,
#> #   Detail_cycle <chr>, Genes <chr>, Gene_description <chr>, Enzyme <chr>,
#> #   KO <chr>, rbims_pathway <chr>, rbims_sub_pathway <chr>, Bin_10 <int>,
#> #   Bin_12 <int>, Bin_56 <int>, Bin_113 <int>, Bin_1 <int>, Bin_2 <int>