Maps KOs to the KEGG database.

Reads a table object created with the read_ko function and maps each KO to the KEGG database.

mapping_ko(tibble_ko=NULL, tibble_interpro=NULL)

Arguments

tibble_ko
a data frame object with four columns containing the KO abundance for each bin. Output of read_ko.

tibble_interpro
a data frame with three columns. The first column refers to contigs. They are expected to indicate in their names the bin name follow by the scaffold name divided by an "_": bin_scaffoldXX. The second column is the KEGG pathway to which such scaffold belong, and the third is the enzymes. Output of read_interpro.

Details

This function is part of a package used for the analysis of bins metabolism.

Examples

mapping_ko(ko_bin_table)

#> # A tibble: 9,882 x 19
#>    Module Module_description     Pathway  Pathway_descripti… Cycle Pathway_cycle
#>    <chr>  <chr>                  <chr>    <chr>              <chr> <chr>        
#>  1 NA     NA                     NA       NA                 NA    NA           
#>  2 NA     NA                     map00564 Glycerophospholip… NA    NA           
#>  3 M00004 Pentose phosphate pat… map00030 Pentose phosphate… NA    NA           
#>  4 M00007 Pentose phosphate pat… map00030 Pentose phosphate… NA    NA           
#>  5 M00004 Pentose phosphate pat… map01100 Metabolic pathways NA    NA           
#>  6 M00007 Pentose phosphate pat… map01100 Metabolic pathways NA    NA           
#>  7 M00004 Pentose phosphate pat… map01110 Biosynthesis of s… NA    NA           
#>  8 M00007 Pentose phosphate pat… map01110 Biosynthesis of s… NA    NA           
#>  9 M00004 Pentose phosphate pat… map01120 Microbial metabol… NA    NA           
#> 10 M00007 Pentose phosphate pat… map01120 Microbial metabol… NA    NA           
#> # … with 9,872 more rows, and 13 more variables: Detail_cycle <chr>,
#> #   Genes <chr>, Gene_description <chr>, Enzyme <chr>, KO <chr>,
#> #   rbims_pathway <chr>, rbims_sub_pathway <chr>, Bin_10 <int>, Bin_12 <int>,
#> #   Bin_56 <int>, Bin_113 <int>, Bin_1 <int>, Bin_2 <int>