Read the output of dbCAN and extract abundance profile

read_dbcan3 calculates the abundance of each Gene within the bins based on the dbCAN output from run_dbcan3 V3.0.6.

read_dbcan3(dbcan_path, write=FALSE, profile=TRUE)

Arguments

dbcan_path

a path where dbCAN output data are. They should have the extension overview.txt and all files in the path are the ones that need to be read. Output data should have 6 columns with the bin names followed by the Genes obtained in every algorithm (HMMER,Hotpep,DIAMOND), column 'Signalp' indcating if a Peptide signal is found and a column '#ofTools" indicating the number of algorithms that found this Gene.

write

a logical value indicating to save the data imported as a formatted table with .tsv extension with a time stamp and it will be located in your current workin directory

profile

a logical value indicating if you want to print a profile or not.

Examples

if (FALSE) { # \dontrun{
read_dbcan3("C:/Users/bins/")
} # }