Metabolism PCA.

Identifies the most important KO pathways or protein domains in the whole database. And print back a profile of the protein domains that have higher contributions.

get_subset_pca(tibble_rbims, cos2_val=NULL, 
analysis=c("KEGG", "PFAM", "INTERPRO"))

Arguments

tibble_rbims
a tibble object, created with the read_interpro, mapping_ko or get_subset_* functions.

cos2_val
a numeric vector from 0 to 1 indicating the proportion of contribution used as cut off. Default is 0.98. See get_pca.

analysis
a character, indicating from which input do you want to get the abundance profile. Valid options are "KEGG", "PFAM" or "INTERPRO".

Details

This function is part of a package used for the analysis of bins metabolism.

Examples

get_subset_pca(ko_bin_mapp, analysis="KEGG")

#> Warning: Contribution of the first dimention is less or equal to 0.97

#> # A tibble: 0 x 19
#> # … with 19 variables: Module <chr>, Module_description <chr>, Pathway <chr>,
#> #   Pathway_description <chr>, Cycle <chr>, Pathway_cycle <chr>,
#> #   Detail_cycle <chr>, Genes <chr>, Gene_description <chr>, Enzyme <chr>,
#> #   KO <chr>, rbims_pathway <chr>, rbims_sub_pathway <chr>, Bin_10 <int>,
#> #   Bin_12 <int>, Bin_56 <int>, Bin_113 <int>, Bin_1 <int>, Bin_2 <int>