Heatmap

Creates a heatmap of KEGG and InterProScan data.

plot_heatmap(tibble_ko, y_axis, analysis=c("KEGG","INTERPRO"), 
data_experiment=NULL, calc=NULL, scale_option=NULL, order_y=NULL, order_x=NULL,
split_y=FALSE, color_pallet=NULL, distance=FALSE)

Arguments

tibble_ko
a data frame object. It could have been created with the mapping_ko or get_subset_* functions. If that is the case you will have to choose which type of analysis do you want to show, "binary", "abundance" or "percentage". You can also include a table with the counts that you want to use.

y_axis
a string. A column name of the tibble_ko of a feature to plot (i.e. KO/Pathways/Modules/PFAM/INTERPRO).

analysis
a character indicating the input file. Valid arguments "KEGG" or "INTERPRO".

data_experiment
optional. a data frame object containing metadata information.

calc
optional. This option is only valid for "KEGG" analysis. Is a character indicating with type of calc should be done to plot the results. Valid values are "Abundance", "Binary", "Percentage", and "None". If you chose none you are spected to use a tibble table obtained from calc_binary or calc_percentage.

scale_option
a character indicating if rows or columns should be scale. Valid options "none", "row" or "column".

order_y
optional. This option is only valid for "KEGG" analysis. Is a column name indicating the annotation if rows. This column name comes from the mapping_ko or get_subset_* object.

order_x
optional. This option is only valid for "KEGG" analysis. Is a column name indicating the annotation if cols. This column name comes from the metadata object.

split_y
optional. This option is only valid for "KEGG" analysis. A logical character indicating if you want to split y axis, based on order_y character.

color_pallet
optional. a character vector of colors to use.

distance
optional. If TRUE it will calculate a distance matrix and show how similar are the different genomes.

Details

This function is part of a package used for the analysis of bins metabolism.

Examples

if (FALSE) {
plot_heatmap(tibble_ko=ko_bin_mapp, y_axis=Pathway,
analysis="INTERPRO", data_experiment = metadata, 
calc="Binary", order_y=Module, order_x=Clades)
}